function_shipment_tracking/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
pandas==2.*
requests==2.*
db-dtypes==1.*  # Required for BigQuery DataFrame conversion